backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.12.0
SQLAlchemy==2.0.27
Werkzeug==3.0.1
marshmallow==3.20.2
python-dateutil==2.8.2
boto3==1.34.69
geopy==2.4.0
requests==2.31.0
pydantic==2.3.0 